// types/type.ts
export enum CursorMode {
  Hidden,
  Chat,
  ReactionSelector,
  Reaction,
}

export type CursorState =
  | {
      mode: CursorMode.Hidden
      message?: string
    }
  | {
      mode: CursorMode.Chat
      message: string
      previousMessage: string | null
    }
  | {
      mode: CursorMode.ReactionSelector
      message?: string
    }
  | {
      mode: CursorMode.Reaction
      reaction: string
      isPressed: boolean
      message?: string
    }

export type Reaction = {
  value: string
  timestamp: number
  point: { x: number; y: number }
}
// ...
export type CursorChatProps = {
  cursor: { x: number; y: number }
  cursorState: CursorState
  setCursorState: (cursorState: CursorState) => void
  updateMyPresence: (
    presence: Partial<{
      cursor: { x: number; y: number }
      cursorColor: string
      message: string
    }>
  ) => void
}